Move CelebrateLaunchModal to common/ for shared use
The modal is now used by both the masterbar launch button and the
dashboard widgets, so it belongs in common/ rather than scoped
to wpcom-dashboard-widgets.
