+++ projects/packages/jetpack-mu-wpcom/src/common/celebrate-launch-modal.scss
