+++ projects/packages/jetpack-mu-wpcom/src/common/confetti-animation.ts
